requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Marshmallow==1.2.1
numpy==2.2.3
marshmallow-sqlalchemy==1.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
PyJWT==2.8.0
Flask-Cors==4.0.1
Werkzeug==3.0.3
Flask-PyMongo==2.3.0
flasgger==0.9.7.1
pandas==2.2.3
Unidecode==1.3.8
